move mill.main.server.Server to mill.server.Server
